move test out of stats/prometheus
