+++ tests/integration/telemetry/api/accesslogs_test.go
@@ -29,7 +29,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package nullvm
+package api
 
 import (
 	"fmt"
