+++ tests/integration/telemetry/api/stats_wasm_tcp_filter_test.go
@@ -15,7 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package telemetryapi
+package api
 
 import (
 	"testing"
