+++ tests/integration/telemetry/envoyfilter/nullvm/stats_filter_test.go
