+++ tests/integration/telemetry/envoyfilter/testdata/istio-mtls-dest-rule.yaml
