+++ tests/integration/telemetry/envoyfilter/testdata/istio-mtls-gateway.yaml
