+++ tests/integration/telemetry/envoyfilter/testdata/istio-mtls-vs.yaml
