+++ tests/integration/telemetry/envoyfilter/wasm/bad_wasm_filter_test.go
