+++ tests/integration/telemetry/envoyfilter/wasm/imagepullpolicy_test.go
