+++ tests/integration/telemetry/envoyfilter/wasm/stats_wasm_filter_test.go
