+++ tests/integration/telemetry/envoyfilter/wasm/testdata/bad-filter.yaml
