+++ tests/integration/telemetry/envoyfilter/wasm/testdata/registry-secret.yaml
