+++ tests/integration/telemetry/envoyfilter/wasm/testdata/wasm-filter.yaml
